+++ app/Catalog/Http/ListCategories/ListCategoriesRequest.php
@@ -19,6 +19,17 @@ class ListCategoriesRequest extends FormRequest
         ];
     }
 
+    /**
+     * @return array<string, bool>
+     */
+    public function inputData(): array
+    {
+        return [
+            'onlyActive' => $this->boolean('only_active', true),
+            'tree'       => $this->boolean('tree', true),
+        ];
+    }
+
     public function authorize(): bool
     {
         return true;